+++ recipes-core/initrdscripts/files/initramfs-init.sh
@@ -156,17 +156,27 @@ pvsn_flash() {
     dmsetup remove /dev/mapper/decrypted-irma6lvm-userdata
 }
 
+emergency_switch() {
+    pending_update=$(/usr/bin/fw_printenv upgrade_available | awk -F'=' '{print $2}')
+    if [ "$pending_update" = "1" ]; then
+        echo "Update pending, let bootcount switch firmware..."; exit 1;   
+    fi
+
+    firmware=$(/usr/bin/fw_printenv firmware | awk -F'=' '{print $2}')
+    if [ "$firmware" -eq 1 ] || [ "$firmware" -eq 0 ]; then
+        firmware=$(( firmware^1 ))
+        /usr/bin/fw_printenv firmware "$firmware"
+        sync
+        echo "Emergency firmware switch to: $firmware"
+    fi
+    exit 1;
+}
+
 mount_pseudo_fs
 
 # populate LVM mapper devices
 vgchange -a y
 vgmknodes
-
-# check if we are in provisioning and need to encrypt the volumes
-echo "Provisioning check..."
-if [ -e "/dev/mapper/irma6lvm-pvsn_rootfs" ]; then
-    pvsn_flash
-fi
 
 echo "Initramfs Bootstrap..."
 parse_cmdline
@@ -180,6 +190,13 @@ then
     exec switch_root ${ROOT_MNT} ${INIT} "${CMDLINE}"
     exit 0
 fi
+
+# check if we are in provisioning and need to encrypt the volumes
+echo "Provisioning check..."
+if [ -e "/dev/mapper/irma6lvm-pvsn_rootfs" ]; then
+    pvsn_flash
+fi
+
 
 KEYSTORE_DEV="/dev/mapper/irma6lvm-keystore"
 KEYSTORE="/mnt/keystore"
@@ -212,7 +229,8 @@ ${MOUNT} ${KEYSTORE_DEV} ${KEYSTORE}
 
 if ! /usr/bin/openssl dgst -sha256 -verify "${PUBKEY}" -signature "${ROOT_HASH_SIGNATURE}" "${ROOT_HASH}"
 then
-    exit 1
+    echo "Root hash signature invalid"
+    emergency_switch
 fi
 RH=$(cat "${ROOT_HASH}")
 
@@ -234,7 +252,11 @@ ${UMOUNT} ${KEYSTORE}
 debug "Opening verity device: ${DECRYPT_ROOT_DEV}"
 veritysetup open ${DECRYPT_ROOT_DEV} ${VERITY_NAME} ${ROOT_HASH_DEV} ${RH}
 
-${MOUNT} ${VERITY_DEV} ${ROOT_MNT} -o ro
+if ! ${MOUNT} ${VERITY_DEV} ${ROOT_MNT} -o ro 
+then
+    echo "Mount root device failed"
+    emergency_switch
+fi
 move_special_devices
 echo "Switching root to verity device"
 exec switch_root ${ROOT_MNT} ${INIT} "${CMDLINE}"