+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [APR's Version Numbering](https://apr.apache.org/ver
 
 ## [2.0.12-dev] (HEAD) - n/a
 ### Added
+- [APC-5579]: factory-reset.sh: Add force option
+
 
 ### Changed