+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [APR's Version Numbering](https://apr.apache.org/ver
 - [APC-5725]: Share factory reset script between R1 and R2
 - [DEVOPS-590]: Add FindAvahi.cmake to avahi and Findswupdate.cmake to swupdate to allow usage in cmake projects
 - [APC-5893]: respect EPC660 specified reset timings
+- [APC-5574]: remove customer folder on factory reset
 
 ### Changed