+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [APR's Version Numbering](https://apr.apache.org/ver
 - [APC-5893]: respect EPC660 specified reset timings
 - [APC-5574]: remove customer folder on factory reset
 - [APC-5792]: Release 2: Default to https in flashall.uuu, swuimage and factory reset script
+- [APC-5942]: Add openssl-bin and download.crt
+
 
 ### Changed